fix #249149
