+++ src/vs/workbench/browser/parts/paneCompositePart.ts
@@ -29,9 +29,8 @@ import { localize } from '../../../nls.js';
 import { CompositeDragAndDropObserver, toggleDropEffect } from '../dnd.js';
 import { EDITOR_DRAG_AND_DROP_BACKGROUND } from '../../common/theme.js';
 import { IPartOptions } from '../part.js';
-import { CompositeMenuActions } from '../actions.js';
 import { IMenuService, MenuId } from '../../../platform/actions/common/actions.js';
-import { ActionsOrientation, prepareActions } from '../../../base/browser/ui/actionbar/actionbar.js';
+import { ActionsOrientation } from '../../../base/browser/ui/actionbar/actionbar.js';
 import { Gesture, EventType as GestureEventType } from '../../../base/browser/touch.js';
 import { StandardMouseEvent } from '../../../base/browser/mouseEvent.js';
 import { IAction, SubmenuAction } from '../../../base/common/actions.js';
@@ -39,7 +38,7 @@ import { Composite } from '../composite.js';
 import { ViewsSubMenu } from './views/viewPaneContainer.js';
 import { getActionBarActions } from '../../../platform/actions/browser/menuEntryActionViewItem.js';
 import { IHoverService } from '../../../platform/hover/browser/hover.js';
-import { HiddenItemStrategy, WorkbenchToolBar } from '../../../platform/actions/browser/toolbar.js';
+import { HiddenItemStrategy, MenuWorkbenchToolBar } from '../../../platform/actions/browser/toolbar.js';
 import { DeferredPromise } from '../../../base/common/async.js';
 
 export enum CompositeBarPosition {
@@ -128,8 +127,8 @@ export abstract class AbstractPaneCompositePart extends CompositePart<PaneCompos
 	private compositeBarPosition: CompositeBarPosition | undefined = undefined;
 	private emptyPaneMessageElement: HTMLElement | undefined;
 
-	private globalToolBar: WorkbenchToolBar | undefined;
-	private readonly globalActions: CompositeMenuActions;
+	private readonly globalActionsMenuId: MenuId;
+	private globalToolBar: MenuWorkbenchToolBar | undefined;
 
 	private blockOpening: DeferredPromise<PaneComposite | undefined> | undefined = undefined;
 	protected contentDimension: Dimension | undefined;
@@ -190,15 +189,13 @@ export abstract class AbstractPaneCompositePart extends CompositePart<PaneCompos
 		);
 
 		this.location = location;
-		this.globalActions = this._register(this.instantiationService.createInstance(CompositeMenuActions, globalActionsMenuId, undefined, undefined));
-
+		this.globalActionsMenuId = globalActionsMenuId;
 		this.registerListeners();
 	}
 
 	private registerListeners(): void {
 		this._register(this.onDidPaneCompositeOpen(composite => this.onDidOpen(composite)));
 		this._register(this.onDidPaneCompositeClose(this.onDidClose, this));
-		this._register(this.globalActions.onDidChange(() => this.updateGlobalToolbarActions()));
 
 		this._register(this.registry.onDidDeregister((viewletDescriptor: PaneCompositeDescriptor) => {
 
@@ -357,17 +354,19 @@ export abstract class AbstractPaneCompositePart extends CompositePart<PaneCompos
 		const globalTitleActionsContainer = titleArea.appendChild($('.global-actions'));
 
 		// Global Actions Toolbar
-		this.globalToolBar = this._register(this.instantiationService.createInstance(WorkbenchToolBar, globalTitleActionsContainer, {
-			actionViewItemProvider: (action, options) => this.actionViewItemProvider(action, options),
-			orientation: ActionsOrientation.HORIZONTAL,
-			getKeyBinding: action => this.keybindingService.lookupKeybinding(action.id),
-			anchorAlignmentProvider: () => this.getTitleAreaDropDownAnchorAlignment(),
-			toggleMenuTitle: localize('moreActions', "More Actions..."),
-			hoverDelegate: this.toolbarHoverDelegate,
-			hiddenItemStrategy: HiddenItemStrategy.NoHide
-		}));
-
-		this.updateGlobalToolbarActions();
+		this.globalToolBar = this._register(this.instantiationService.createInstance(MenuWorkbenchToolBar,
+			globalTitleActionsContainer,
+			this.globalActionsMenuId,
+			{
+				actionViewItemProvider: (action, options) => this.actionViewItemProvider(action, options),
+				orientation: ActionsOrientation.HORIZONTAL,
+				getKeyBinding: action => this.keybindingService.lookupKeybinding(action.id),
+				anchorAlignmentProvider: () => this.getTitleAreaDropDownAnchorAlignment(),
+				toggleMenuTitle: localize('moreActions', "More Actions..."),
+				hoverDelegate: this.toolbarHoverDelegate,
+				hiddenItemStrategy: HiddenItemStrategy.NoHide
+			}
+		));
 
 		return titleArea;
 	}
@@ -626,12 +625,6 @@ export abstract class AbstractPaneCompositePart extends CompositePart<PaneCompos
 		if (visible) {
 			this.titleLabel?.updateTitle('', '');
 		}
-	}
-
-	private updateGlobalToolbarActions(): void {
-		const primaryActions = this.globalActions.getPrimaryActions();
-		const secondaryActions = this.globalActions.getSecondaryActions();
-		this.globalToolBar?.setActions(prepareActions(primaryActions), prepareActions(secondaryActions));
 	}
 
 	protected getToolbarWidth(): number {
